navbar skeletal structure and logout logic and button complete
