+++ src/pages/Home.js
@@ -52,6 +52,8 @@ function Home({ handleLogOut }) {
 
     const navigate = useNavigate();
 
+    // WRITE OUT useEffect function to check for displayName, if don't have, redirect to CreateProfile.js
+
     useEffect(() => {
         const postsRef = dbRef(database, DB_POSTS_KEY);
 
@@ -128,10 +130,10 @@ function Home({ handleLogOut }) {
         setSelectedPost(updatedPost);
     }, [posts]);
 
-    // const { isLoaded } = useJsApiLoader({
-    //     googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAP_API_KEY,
-    //     libraries: libraries,
-    // });
+    const { isLoaded } = useJsApiLoader({
+        googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAP_API_KEY,
+        libraries: libraries,
+    });
 
     // const coords = useMemo(() => ({ lat: 1.3521, lng: 103.8198 }), []);
     const mapRef = useRef();
@@ -187,13 +189,9 @@ function Home({ handleLogOut }) {
         setShowBazaarForm(false);
     };
 
-    const handleLogOutAndNavigate = () => {
-        handleLogOut().then(() => navigate("/"));
-    };
-
-    // if (!isLoaded) {
-    //     return <Typography variant="h1">Loading...</Typography>;
-    // }
+    if (!isLoaded) {
+        return <Typography variant="h1">Loading...</Typography>;
+    }
 
     return (
         <Box
@@ -205,14 +203,6 @@ function Home({ handleLogOut }) {
         >
             <SidebarWrapper>
                 Welcome back {displayName}
-                <Button
-                    variant="contained"
-                    sx={{ width: "50%", m: 1 }}
-                    onClick={handleLogOutAndNavigate}
-                    type="button"
-                >
-                    Log Out
-                </Button>
                 <Typography variant="body1">
                     Type in a place or click on the map to get started
                 </Typography>
@@ -257,7 +247,7 @@ function Home({ handleLogOut }) {
                 )}
             </SidebarWrapper>
 
-            {/* <GoogleMap
+            <GoogleMap
                 center={coords}
                 zoom={13}
                 mapContainerStyle={{ width: "100%", height: "100%" }}
@@ -278,7 +268,7 @@ function Home({ handleLogOut }) {
                             }}
                         />
                     ))}
-            </GoogleMap> */}
+            </GoogleMap>
         </Box>
     );
 }
